fix: co-locate player types
