+++ game/src/routes/game/+page.svelte
@@ -1,11 +1,19 @@
 <script lang="ts">
 	import { browser } from '$app/environment';
 	import { ATLASES, getRandomAtlas, type Atlas } from '$lib/atlases';
-	import { type Game, getRandomValue } from '$lib/helpers';
-	import { getRounds, getDecoyRound } from '$lib/rounds';
+	import { getRandomValue } from '$lib/helpers';
+	import { getRank } from '$lib/player';
+	import { getRounds, getDecoyRound, type Round } from '$lib/rounds';
 	import { playerStore } from '$lib/stores/player';
 	import { SUSPECTS, type Suspect } from '$lib/suspects';
 	import { format } from 'date-fns';
+
+	interface Game {
+		currentTime: Date;
+		stolenTreasure: string;
+		suspect: Suspect;
+		rounds: Round[];
+	}
 
 	// If there is no user profile, redirect to the player page
 	if ($playerStore === null && browser) window.location.href = '/player';
@@ -72,7 +80,6 @@
 <form>
 	<fieldset>
 		<legend>Debug controls</legend>
-		<p>$playerStore — Name: <u>{$playerStore?.name}</u> - Score: <u>{$playerStore?.score}</u></p>
 		<p>
 			<strong> rounds: </strong>
 			{#each game.rounds as round, i}
@@ -91,6 +98,10 @@
 		>
 	</fieldset>
 </form>
+
+{#if $playerStore}
+	<p>Name: <u>{$playerStore.name}</u> - Rank: <u>{getRank($playerStore.score)}</u></p>
+{/if}
 
 <p>The stolen item is <u>{game.stolenTreasure}</u></p>
 <p>The suspect sex is <u>{game.suspect.sex}</u></p>
